requirements.txt
streamlit
openai-whisper
jiwer
soundfile
